app/requirements.txt
sqlalchemy
python-dotenv
flask
authlib
faker
requests
pymysql
pandas